+++ src/LIN3S/AdminBundle/DependencyInjection/CompilerPass/ListFiltersTypePass.php
@@ -20,7 +20,7 @@ use Symfony\Component\DependencyInjection\Reference;
  *
  * @author Gorka Laucirica <gorka.lauzirka@gmail.com>
  */
-class ListFiltersTypePass implements CompilerPassInterface
+final class ListFiltersTypePass implements CompilerPassInterface
 {
     /**
      * {@inheritdoc}